requirements.txt
gymnasium==1.2.0
stable_baselines==2.7.0
numpy==1.26.2
torch==2.7.1
